pyproject.toml
[project]
name = "paleo-bench"
version = "0.1.0"
description = "Handwritten text recognition benchmark for LLM vision models"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "litellm>=1.40",
    "python-dotenv>=1.0",
    "jiwer>=3.0",
    "rapidfuzz>=3.0",
    "pillow>=12.1.1",
]
